+++ SwiftCompilerSources/Sources/Optimizer/InstructionSimplification/CMakeLists.txt
@@ -10,7 +10,7 @@ swift_compiler_sources(Optimizer
   SimplifyAllocRefDynamic.swift
   SimplifyAllocStack.swift
   SimplifyApply.swift
-  SimplifyBeginAndLoadBorrow.swift
+  SimplifyBeginBorrow.swift
   SimplifyBeginCOWMutation.swift
   SimplifyBranch.swift
   SimplifyBuiltin.swift
@@ -31,6 +31,7 @@ swift_compiler_sources(Optimizer
   SimplifyInitEnumDataAddr.swift
   SimplifyKeyPath.swift
   SimplifyLoad.swift
+  SimplifyLoadBorrow.swift
   SimplifyMarkDependence.swift
   SimplifyMisc.swift
   SimplifyPartialApply.swift
